Update SearchRequestAdapter.js
This small fix would allow "empty" queries to proceed without any exceptions from this adapter.

For example: If you do not add a search box instantsearch will not set a query and this adapter would fail.

Also it solves this issue #43 that I had. For some reason instantsearch triggers an empty request in my setup and then comes back with an error which results in absolute failure. Also resolves #44 

Would be great if we could merge this!